Bump minimum Peribot version to 0.6.6
Ensure that we are using at least Peribot 0.6.6. This version includes
some modifications to enhance the utility of the Peribot::Bot#use
method. Specifically, all Service subclasses now support a register_into
method, and Processor subclasses are expected to implement a
register_into method as appropriate. Changes will be made to
Peribot::GroupMe to give an example of how this can be done properly.