Remove all instance doubles of Peribot::Bot
Perhaps due to the relative difficulty of setting up previous versions
of Peribot with directory-based configuration and storage, I was pretty
heavily relying on RSpec instance doubles to duplicate the functionality
of Peribot::Bot when it was used. Obviously, this is horrifying and
embarassing.

I'm now using temporary files and block-based configuration to create a
real Bot instance to use during all tests.