Update user sender for new message format
Ensure that the user sender works properly with the new message format.
The behavior has been changed in a similar fasion to the user like
sender. Instead of stopping processing on invalid messages, they are
returned for further processing. Again, this should ease the transition
to Peribot 0.8's new sender architecture.