Use Peribot 0.10's filter chain for bot filtering
Peribot 0.10 introduces a separation between ingress filtering and
preprocessing of messages, to better support certain use cases. Because
the bot filtering provided by Peribot::GroupMe acts in this capacity, it
is changed from a preprocessor to a filter for Peribot 0.10. As part of
this the class has been renamed to better reflect its purpose.