Allow case of empty text from GroupMe push
In some cases, it is possible for GroupMe to send a message that does
not contain *any* text at all. Peribot::GroupMe would simply pass
through a nil text value to Peribot, which Peribot::Service would choke
on due to a consistency check on required fields (text, service, group).

With this change, Peribot should now be able to successfully handle such
messages, since we just use an empty string for text rather than nil.
The point is that this is still a string rather than something else.